Rename schemas/chat_notifications -> chat_notification
