+++ imports/lib/models/chat_notifications.ts
@@ -1,8 +1,8 @@
-import ChatNotificationsSchema, { ChatNotificationType } from '../schemas/chat_notifications';
+import ChatNotificationSchema, { ChatNotificationType } from '../schemas/chat_notification';
 import Base from './base';
 
 const ChatNotifications = new Base<ChatNotificationType>('chatnotifications');
-ChatNotifications.attachSchema(ChatNotificationsSchema);
+ChatNotifications.attachSchema(ChatNotificationSchema);
 // No publish here -- we do a custom publish for notifications matching the target user.
 
 export default ChatNotifications;
