+++ imports/server/chat-notifications.ts
@@ -5,7 +5,7 @@ import Flags from '../flags';
 import ChatNotifications from '../lib/models/chat_notifications';
 import Hunts from '../lib/models/hunts';
 import Puzzles from '../lib/models/puzzles';
-import { ChatNotificationType } from '../lib/schemas/chat_notifications';
+import { ChatNotificationType } from '../lib/schemas/chat_notification';
 import { HuntType } from '../lib/schemas/hunts';
 import { PuzzleType } from '../lib/schemas/puzzles';
 import RefCountedObserverMap from './refcounted-observer-map';
